+++ packages/framework/src/Foundation/Hyperlinks.php
@@ -31,7 +31,7 @@ class Hyperlinks
      */
     public function formatLink(string $destination): string
     {
-        if (config('site.pretty_urls', false) === true) {
+        if (config('hyde.pretty_urls', false) === true) {
             if (str_ends_with($destination, '.html')) {
                 if ($destination === 'index.html') {
                     return '/';
@@ -108,7 +108,7 @@ class Hyperlinks
      */
     public function hasSiteUrl(): bool
     {
-        return ! blank(config('site.url'));
+        return ! blank(config('hyde.url'));
     }
 
     /**
@@ -123,7 +123,7 @@ class Hyperlinks
         $path = $this->formatLink(trim($path, '/'));
 
         if ($this->hasSiteUrl()) {
-            return rtrim(rtrim((string) config('site.url'), '/')."/$path", '/');
+            return rtrim(rtrim((string) config('hyde.url'), '/')."/$path", '/');
         }
 
         throw new BaseUrlNotSetException();